Moved UserImage model to separate file
